Set 'name' option to produce legal WiX identifiers
It seems that the nuget identifier, generated from the 'name' property
of package.json, can cause trouble with WiX.

We can override the value from the options. Note that this is not part
of the official API of electron-winstaller, though.

Another solution would be to change the 'name' property in package.json,
but that seems too unreasonable.