Add throttling to sampling period
Because the first generation Senso Flex prototype sends at its own
sampling rate, this may differ from the poll-based rate we set for the
second generation.

This might mess with assumptions in clients based upon that sampling
rate, therefore we throttle the sends to at most the rate of the
poll-based sampling.